requirements.txt
aiofiles==23.2.1
aiogram==2.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
black==24.4.2
certifi==2024.6.2
click==8.1.7
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isort==5.13.2
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
typing_extensions==4.12.2
yarl==1.9.4
